+++ contracts/core/IController.sol
@@ -3,15 +3,15 @@
 pragma solidity ^0.8.11;
 
 interface IController {
-    function protocolFeeBips() external view returns (uint256);
+    function protocolFeeBips() external view returns (uint64);
 
     function treasuryAddress() external view returns (address);
 
-    function distributitorFeeBips() external view returns (uint256);
+    function distributitorFeeBips() external view returns (uint64);
 
-    function setProtocolFee(uint256 bips) external;
+    function setProtocolFee(uint64 bips) external;
 
     function setTreasury(address target) external;
 
-    function setDistributorFee(uint256 bips) external;
+    function setDistributorFee(uint64 bips) external;
 }
